Hzero authority

See merge request hzero-agile/agile-service!507

+++ src/main/java/io/choerodon/agile/app/service/impl/IssueAttachmentServiceImpl.java
@@ -129,6 +129,9 @@ public class IssueAttachmentServiceImpl implements IssueAttachmentService {
         if (issueAttachmentDTO == null) {
             throw new CommonException("error.attachment.get");
         }
+        if (!issueAttachmentDTO.getProjectId().equals(projectId)) {
+            throw new CommonException("error.project.id.does.not.correspond");
+        }
         Boolean result = iIssueAttachmentService.deleteBase(issueAttachmentDTO.getAttachmentId());
         BaseFieldUtil.updateIssueLastUpdateInfo(issueAttachmentDTO.getIssueId(), issueAttachmentDTO.getProjectId());
         String url = null;
